Add SM90 to matmul ops.  Move matmul functions to composite ops.  Add tests.
