+++ csrc/ops/composite.cpp
@@ -5,9 +5,9 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 // clang-format on
+#include <ATen/cuda/CUDAContext.h>
 #include <ir_builder.h>
-#include <ops/arith.h>
-#include <ops/composite.h>
+#include <ops/all_ops.h>
 #include <transform_view.h>
 
 namespace nvfuser {
@@ -51,6 +51,72 @@ TensorView* dropout_backward(TensorView* dy, TensorView* mask, Val* scale) {
   auto dx = mul(grad_mask, scale);
 
   return dx;
+}
+
+// Matmul functions are temporary internal functions for testing purposes only
+TensorView* _matmul_nn(TensorView* a, TensorView* b) {
+  TORCH_CHECK(
+      a->nDims() == 2 && b->nDims() == 2, "Only 2-D Tensors are supported!");
+  const auto device_prop = at::cuda::getCurrentDeviceProperties();
+  bool is_turing_or_later = ((device_prop->major == 7) && (device_prop->minor >= 5)) || (device_prop->major >= 8);
+  TensorView *tv0b = nullptr, *tv1b = nullptr, *tv2 = nullptr;
+  if (is_turing_or_later) {
+    auto tv0t = transpose(a, 0, 1);
+    tv0b = broadcast(tv0t, {false, true, false});
+    tv1b = broadcast(b, {true, false, false});
+    tv2 = fusedMultiplySum(tv0b, tv1b, {2});
+    return tv2;
+  } else {
+    tv0b = broadcast(b, {false, false, true});
+    tv1b = broadcast(a, {true, false, false});
+    tv2 = fusedMultiplySum(tv0b, tv1b, {1});
+    return transpose(tv2, 0, 1);
+  }
+}
+TensorView* _matmul_nt(TensorView* a, TensorView* b) {
+  TORCH_CHECK(
+      a->nDims() == 2 && b->nDims() == 2, "Only 2-D Tensors are supported!");
+  const auto device_prop = at::cuda::getCurrentDeviceProperties();
+  bool is_turing_or_later = ((device_prop->major == 7) && (device_prop->minor >= 5)) || (device_prop->major >= 8);
+  TensorView *tv0b = nullptr, *tv1b = nullptr, *tv2 = nullptr;
+  if (is_turing_or_later) {
+    auto tv0t = transpose(a, 0, 1);
+    auto tv1t = transpose(b, 0, 1);
+    tv0b = broadcast(tv0t, {false, true, false});
+    tv1b = broadcast(tv1t, {true, false, false});
+    tv2 = fusedMultiplySum(tv0b, tv1b, {2});
+  } else {
+    tv0b = broadcast(a, {false, false, true});
+    tv1b = broadcast(b, {false, true, false});
+    tv2 = fusedMultiplySum(tv0b, tv1b, {0});
+  }
+  return tv2;
+}
+TensorView* _matmul_tn(TensorView* a, TensorView* b) {
+  TORCH_CHECK(
+      a->nDims() == 2 && b->nDims() == 2, "Only 2-D Tensors are supported!");
+  auto tv0b = broadcast(a, {false, true, false});
+  auto tv1b = broadcast(b, {true, false, false});
+  auto tv2 = fusedMultiplySum(tv0b, tv1b, {2});
+  return tv2;
+}
+TensorView* _matmul_tt(TensorView* a, TensorView* b) {
+  TORCH_CHECK(
+      a->nDims() == 2 && b->nDims() == 2, "Only 2-D Tensors are supported!");
+  const auto device_prop = at::cuda::getCurrentDeviceProperties();
+  bool is_turing_or_later = ((device_prop->major == 7) && (device_prop->minor >= 5)) || (device_prop->major >= 8);
+  TensorView *tv0b = nullptr, *tv1b = nullptr, *tv2 = nullptr;
+  if (is_turing_or_later) {
+    auto tv1t = transpose(b, 0, 1);
+    tv0b = broadcast(a, {false, true, false});
+    tv1b = broadcast(tv1t, {true, false, false});
+    tv2 = fusedMultiplySum(tv0b, tv1b, {2});
+  } else {
+    tv0b = broadcast(a, {false, false, true});
+    tv1b = broadcast(b, {true, false, false});
+    tv2 = fusedMultiplySum(tv0b, tv1b, {1});
+  }
+  return tv2;
 }
 
 LstmResult lstm(
