+++ csrc/serde/fusion_record_serde.cpp
@@ -744,6 +744,11 @@ void RecordFunctorFactory::setupFunctionMaps() {
       ("ops." op_str), static_cast<TensorView* (*)(TensorView*)>(op_name)); \
   unary_val.emplace(("ops." op_str), static_cast<Val* (*)(Val*)>(op_name));
 
+#define NVFUSER_BINARY_TV_ONLY_OP(op_str, op_name)                      \
+  binary_tv.emplace(                                                    \
+      ("ops." op_str),                                                  \
+      static_cast<TensorView* (*)(TensorView*, TensorView*)>(op_name)); \
+
 #define NVFUSER_BINARY_TV_OP(op_str, op_name)                           \
   binary_tv.emplace(                                                    \
       ("ops." op_str),                                                  \
@@ -884,6 +889,11 @@ void RecordFunctorFactory::setupFunctionMaps() {
   NVFUSER_UNARY_TV_OP("isreal", isreal)
   NVFUSER_UNARY_TV_OP("real", real)
   NVFUSER_UNARY_TV_OP("imag", imag)
+  
+  NVFUSER_BINARY_TV_ONLY_OP("_matmul_nn", _matmul_nn)
+  NVFUSER_BINARY_TV_ONLY_OP("_matmul_nt", _matmul_nt)
+  NVFUSER_BINARY_TV_ONLY_OP("_matmul_tn", _matmul_tn)
+  NVFUSER_BINARY_TV_ONLY_OP("_matmul_tt", _matmul_tt)
 
   NVFUSER_BINARY_TV_OP("add", add)
   NVFUSER_BINARY_TV_OP("atan2", atan2)
